% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genefunnel.R
\name{genefunnel}
\alias{genefunnel}
\title{GeneFunnel scoring algorithm}
\usage{
genefunnel(mat, gene_sets, BPPARAM = BiocParallel::bpparam())
}
\arguments{
\item{mat}{A base numeric/integer matrix or numeric \code{Matrix} object, with
features in rows and samples in columns. Row names must be unique.}

\item{gene_sets}{A named list of character vectors containing feature
identifiers.}

\item{BPPARAM}{A \code{BiocParallelParam} backend.}
}
\value{
A numeric matrix of gene set scores (gene sets x samples).
}
\description{
This function computes GeneFunnel scores for each sample and gene set, based
on a matrix of feature expression and a list of gene sets.
}
