% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genefunnel.R
\name{genefunnel}
\alias{genefunnel}
\title{GeneFunnel scoring algorithm}
\usage{
genefunnel(mat, gene_sets, BPPARAM = bpparam())
}
\arguments{
\item{mat}{A (sparse) matrix of feature expression (features x samples).}

\item{gene_sets}{A named list of character vectors representing gene sets.}

\item{BPPARAM}{A BiocParallel backend.}
}
\value{
A numeric matrix of gene set scores (gene sets x samples).
}
\description{
This function computes GeneFunnel scores for each sample and gene set, based
on a matrix of feature expression and a list of gene sets.
}
